responsive staking pool page
